common_stat

Module cgroups

Structs

CgroupsMetricsCollector

A collector that collects cgroups metrics.

Constants

CGROUP_UNIFIED_MOUNTPOINT 🔒

CPU_MAX_FILE_CGROUP_V2 🔒

CPU_PERIOD_FILE_CGROUP_V1 🔒

CPU_QUOTA_FILE_CGROUP_V1 🔒

CPU_USAGE_FILE_CGROUP_V2 🔒

MAX_MEMORY_IN_BYTES 🔒

MAX_VALUE_CGROUP_V2 🔒

MEMORY_MAX_FILE_CGROUP_V1 🔒

MEMORY_MAX_FILE_CGROUP_V2 🔒

MEMORY_USAGE_FILE_CGROUP_V2 🔒

Functions

calculate_cpu_usage 🔒

get_cgroup_v2_cpu_limit 🔒

get_cpu_limit_from_cgroups

Get the limit of cpu in millicores from cgroups filesystem.

get_cpu_usage_from_cgroups

Get the usage of cpu in millicores from cgroups filesystem.

get_memory_limit_from_cgroups

Get the limit of memory in bytes from cgroups filesystem.

get_memory_usage_from_cgroups

Get the usage of memory in bytes from cgroups filesystem.

is_cgroup_v2 🔒

read_value_from_file 🔒